+++ src/components/CreateCourse/components/AuthorItem/AuthorItem.tsx
@@ -1,17 +1,22 @@
 import React, { FC, useContext } from 'react';
 import Button from 'src/components/common/Button/Button';
-import { CoursesContext } from 'src/components/context/courseContext';
-import { Author } from '../Authors/Authors';
-
-interface AuthorItemProps {
-	author: Author;
-}
+import { CoursesContext } from 'src/context/courseContext';
+import { AuthorItemProps } from 'src/types/common/types';
 
 const AuthorItem: FC<AuthorItemProps> = ({ author }) => {
 	const { setChosenAuthors, chosenAuthors } = useContext(CoursesContext);
 
 	const handleClick = () => {
-		//setChosenAuthors(author);
+		// Check if the author is already in the chosenAuthors array
+		const isAuthorAlreadyChosen = chosenAuthors.some(
+			(chosenAuthor) => chosenAuthor.id === author.id
+		);
+
+		if (!isAuthorAlreadyChosen) {
+			// If the author is not already chosen, add it to the chosenAuthors array
+			const updatedChosenAuthors = [...chosenAuthors, author];
+			setChosenAuthors(updatedChosenAuthors);
+		}
 	};
 
 	return (
